+++ RELEASE_NOTES.md
@@ -1,5 +1,11 @@
 # Change log
 
+## 0.14.1 (2021-06-08)
+
+* Allows automatic coersion of `read_*` return values from `String` to other primitive types
+* Allows automatic coercion of optional-typed expressions to String within placeholders
+* Fixes handling of relative imports that are not in a subdirectory of the main WDL's parent directory
+
 ## 0.14.0 (2021-05-27)
 
 * Adds `fix` tool to automatically fix some non-compliant WDL syntax